+++ main.py
@@ -1,6 +1,7 @@
 #!/usr/bin/env python3
-from src.config import AppConfig
+from src.configs import AppConfig
 from src.auth import authenticate
+from src.core.app import App
 
 
 def main():
@@ -9,18 +10,12 @@ def main():
     """
     # Load configs
     # config = AppConfig.default()
-    # config = AppConfig.development()
+    config = AppConfig.development()
     # config = AppConfig.production()
-    config = AppConfig.debug_mode()
+    # config = AppConfig.debug_mode()
 
-    # Setup Auth
     authenticate()
-
-    # Create and run app
-    # TODO: Replace with proper ChatApp when ready
-    from src.core.temp_app import TempApp
-
-    app = TempApp(config)
+    app = App(config)
     app.run()