+++ src/configs/__init__.py
@@ -1,0 +1,19 @@
+"""
+Core application configuration module.
+
+This package contains the main application configs meant to manage
+the entire chat application and limitations.
+"""
+
+from .models import SupportedModel
+from .chat import ChatConfig
+from .app import AppConfig, ModelConfig
+from .debug import DebugConfig
+
+__all__ = [
+    "SupportedModel",
+    "ModelConfig",
+    "ChatConfig",
+    "AppConfig",
+    "DebugConfig",
+]