+++ src/factories/__init__.py
@@ -3,9 +3,12 @@ Factory classes for creating components from configuration.
 """
 
 from .model import ModelFactory, create_mistral, create_dialogpt
+from .storage import StorageFactory, create_chromadb_store
 
 __all__ = [
     "ModelFactory",
     "create_mistral", 
-    "create_dialogpt"
+    "create_dialogpt",
+    "StorageFactory",
+    "create_chromadb_store"
 ]