+++ src/storage/__init__.py
@@ -1,0 +1,12 @@
+"""
+Storage components for persistent data management.
+"""
+
+from .vector_store_interface import VectorStoreInterface, VectorStoreConfig
+from .chromadb_store import ChromaDBStore
+
+__all__ = [
+    "VectorStoreInterface",
+    "VectorStoreConfig", 
+    "ChromaDBStore"
+]