+++ src/utils/__init__.py
@@ -1,11 +1,12 @@
-# src/utils/__init__.py
 """
 Utility functions and classes
 """
 
 from .logger import Logger, create_logger
+from .graceful_session_manager import GracefulSessionManager
 
 __all__ = [
     "Logger",
-    "create_logger"
+    "create_logger",
+    "GracefulSessionManager",
 ]